Add files via upload
Our system and our currency the FCI will provide;✔ A fast and secure way to carry out global transactions.✔ A way of reducing transactions costs greatly globally.✔ Reduce major fluctuation risks.✔ Help to boost trade globally.✔ Help all nations generate huge savings and the needed new money to solve their problems.✔ A huge helper in fighting hyperinflation and normal inflation.✔ Will eliminate the Triffin Dilemma and the Beggar-thy-Neighbor problems so that a country will stand by itself and only its monetary skills will determine success or failure rather than be derailed by external forces.✔ Will not take away the sovereignty of any country which is the only way to be self-sufficient and the road to wealth, with the current system whenever a country uses another’s as a reserve currency it automatically loses this sovereignty and the ability to control its fiscal planning e.g. by devaluing the currency to increase exports. It relies on adjustments by the country whose reserve currency is in use and therefore technically under it.